feat(rTrees): integrate rTrees and move integrations to their own folder
