+++ src/App.css
@@ -207,6 +207,19 @@ h1 {
   padding-bottom: 7px;
 }
 
+.click-zone.g {
+  left: calc(var(--metagame-left) + 360px * var(--metagame-scale-ratio));
+  top: calc(var(--metagame-top) + 295px * var(--metagame-scale-ratio));
+  width: calc(140px* var(--metagame-scale-ratio));
+  height: calc(25px* var(--metagame-scale-ratio));
+  background-color: transparent;
+  background-image: unset;
+  border: none;
+  padding-left: 5px;
+  padding-top: 10px;
+  padding-bottom: 7px;
+}
+
 .click-zone:hover {
   border: solid 0.5px greenyellow;;
 }
