--- src/components/LoftRadioInstance.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-function LoftRadioInstance({width, height}) {
-    return (
-        <iframe title="Interspace loft.radio" src="https://loft.radio/" style={{height: `${height}px`, width: `${width}px`, border: '0px'}}></iframe>
-    )
-}
-
-export default LoftRadioInstance;
